@@ -7,7 +7,11 @@ import {
   getCurrentUserDetails,
   getUserDetailsById,
   addCourseToUser,
-  googleSignInOrSignUp
+  googleSignInOrSignUp,
+  githubSignInOrSignUp,
+  addDetailsToUser,
+  addUsername,
+  getAllUser,
 } from "../controllers/userControllers";
 import checkAuth from "../middleware/checkAuth";
 
@@ -17,8 +21,12 @@ router.route("/register").post(registerUser);
 router.route("/login").post(loginUser);
 router.route("/verify/:token").get(verifyUser);
 router.get("/me", checkAuth, getCurrentUserDetails); // get the user details of the current user
-router.get("/get/:userId", checkAuth, getUserDetailsById); // get the user details of a specific user
+// router.get("/get/:userId", checkAuth, getUserDetailsById); // get the user details of a specific user
 router.post("/addcourse", checkAuth, addCourseToUser); // add a course to the current user
 router.post("/google", googleSignInOrSignUp); // sign in or sign up using google
+router.post("/github", githubSignInOrSignUp); // sign in or sign up using github
+router.post("/addDetails", addDetailsToUser); // add details to the current user
+router.post("/addusername", addUsername); // change the username of the current user
+router.get('/all', getAllUser);
 
 export default router;
